Move views to viz folder
