+++ js/viz/views/templates/association-view.js
