+++ js/viz/views/templates/list.js
