+++ js/viz/views/templates/tree.js
